+++ packages/ui/src/button/stories/index.story.tsx
@@ -5,6 +5,7 @@ import { Button } from '../index';
 
 const meta: Meta< typeof Button > = {
 	title: 'Design System/Components/Button',
+	synonyms: [ 'action', 'cta', 'submit' ],
 	component: Button,
 	subcomponents: {
 		'Button.Icon': Button.Icon,