+++ packages/ui/src/form/input-control/stories/index.story.tsx
@@ -16,6 +16,7 @@ import {
 
 const meta: Meta< typeof InputControl > = {
 	title: 'Design System/Components/Form/InputControl',
+	synonyms: [ 'field', 'form', 'input', 'label', 'text' ],
 	component: InputControl,
 	argTypes: {
 		defaultValue: { control: false },