+++ packages/ui/src/stack/stories/index.story.tsx
@@ -4,6 +4,7 @@ import { Stack } from '../index';
 const meta: Meta< typeof Stack > = {
 	tags: [ 'manifest' ],
 	title: 'Design System/Components/Stack',
+	synonyms: [ 'flex', 'layout', 'spacing' ],
 	component: Stack,
 	parameters: {
 		componentStatus: {